YearCalendar Cleanup
